client -> vm execution: removed headerfetcher -> blockfetcher inheritance for a clear synchronizer object association
